Strong scaling test works! Parallel Overhead data not so good. Weak scaling test in the making

+++ Assignements/run_exe.sh
@@ -15,48 +15,66 @@ run_parallel(){
 }
 
 #Pass filename_p as first argument, string to grep as second argument, csv name as third
+#Our csv will have these columns: Rank,P,Time,N
 result_files_to_csv(){
 
-	WALLTIME=$(grep $2 $1 | cut -d":" -f2)
-	#USRTIME=$(grep $4 $1 | cut)
 	P=$(echo $1 | cut -d"_" -f2)
 	N=$(echo $1 | cut -d"_" -f3)
-	echo "${P};${OUR_TIME};${N}" >> $3
-
+	#In case we are extracting walltime from serial app
+	if (( $P == 1 ))
+	then
+		WALLTIME=$(grep "walltime" $1 | cut -d":" -f2)
+		echo "${P};${P};${WALLTIME};${N}" >> $2
+	else
+		for ((rank=0;rank<$P;rank++))
+		do
+			if (( $rank == 0 ))
+			then
+				WALLTIME=$(grep "walltime" $1 | grep "master" | cut -d":" -f2)
+			else
+				WALLTIME=$(grep "walltime" $1 | grep -w "processor ${rank}" | cut -d":" -f2)
+			fi
+			echo "${rank};${P};${WALLTIME};${N}" >> $2
+		done
+	fi
 }
 
-#To use on ulysses
-#ASSIGNMENT_PATH="/home/fbrand/FHPC"
-#cd $ASSIGNMENT_PATH
-
 CURR_PROGRAM=$1
+SCALING_TEST=$2
+
 if [[ $CURR_PROGRAM == *"mpi"* ]]
 then
 	module load openmpi
 fi
 
+P_VALUES=$(grep P_VALUES_{$SCALING_TEST} values.cfg | cut -d"=" -f2)
+N_VALUES=$(grep N_VALUES_{$SCALING_TEST} values.cfg | cut -d"=" -f2)
+RESULT_FILE=$(echo $CURR_PROGRAM | cut -d"." -f1)_{$SCALING_TEST}.res
+CSV_FILE=$(echo $CURR_PROGRAM | cut -d"." -f1)_{$SCALING_TEST}.csv
+
 #echo $CURR_PROGRAM
-
-P_VALUES=$(grep P_VALUES values.cfg | cut -d"=" -f2)
-N_VALUES=$(grep N_VALUES values.cfg | cut -d"=" -f2)
-RESULT_FILE=$(echo $CURR_PROGRAM | cut -d"." -f1).res
-CSV_FILE=$(echo $CURR_PROGRAM | cut -d"." -f1).csv
-
-#echo $RESULT_FILE
-#echo $P_VALUES
-#echo $N_VALUES
-
-for n_val in $N_VALUES
-do
-	for p_val in $P_VALUES
+if [[ $SCALING_TEST == "STRONG" ]]
+then
+	for n_val in $N_VALUES
 	do
-		if [[ $CURR_PROGRAM == *"mpi"* ]]
-		then
-			run_parallel $p_val $n_val $RESULT_FILE $CURR_PROGRAM
-		else
-			run_serial $n_val ${RESULT_FILE}_${p_val}_${n_val} $CURR_PROGRAM
-			result_files_to_csv ${RESULT_FILE}_${p_val}_${n_val} "walltime|usr" $CSV_FILE
-			rm ${RESULT_FILE}_${p_val}_${n_val}
-		fi
+		for p_val in $P_VALUES
+		do
+			if [[ $CURR_PROGRAM == *"mpi"* ]]
+			then
+				run_parallel $p_val $n_val ${RESULT_FILE}_${p_val}_${n_val} $CURR_PROGRAM
+				result_files_to_csv ${RESULT_FILE}_${p_val}_${n_val} $CSV_FILE
+				rm ${RESULT_FILE}_${p_val}_${n_val}
+			else
+				run_serial $n_val ${RESULT_FILE}_${p_val}_${n_val} $CURR_PROGRAM
+				result_files_to_csv ${RESULT_FILE}_${p_val}_${n_val} $CSV_FILE
+				rm ${RESULT_FILE}_${p_val}_${n_val}
+			fi
+		done
 	done
-done
+else
+	for p_val in $P_VALUES
+
+#Use millisecond instead of second
+awk 'BEGIN { FS=";" } {printf"%d;%d;%.7f;%d\n", $1, $2, $3*1000, $4}' $CSV_FILE > ${CSV_FILE}_out
+#rm $CSV_FILE
+#mv ${CSV_FILE}_out $CSV_FILE
